# Speed up epoch transition using persistent vector