--- a/packages/lodestar-beacon-state-transition/src/fast/block/processAttesterSlashing.ts
+++ b/packages/lodestar-beacon-state-transition/src/fast/block/processAttesterSlashing.ts
@@ -1,13 +1,13 @@
-import {AttesterSlashing, BeaconState, ValidatorIndex} from "@chainsafe/lodestar-types";
+import {AttesterSlashing, ValidatorIndex} from "@chainsafe/lodestar-types";
 
 import {isSlashableValidator, isSlashableAttestationData} from "../../util";
-import {EpochContext} from "../util";
+import {EpochContext, CachedValidatorsBeaconState} from "../util";
 import {slashValidator} from "./slashValidator";
 import {isValidIndexedAttestation} from "./isValidIndexedAttestation";
 
 export function processAttesterSlashing(
   epochCtx: EpochContext,
-  state: BeaconState,
+  state: CachedValidatorsBeaconState,
   attesterSlashing: AttesterSlashing,
   verifySignatures = true
 ): void {