--- a/packages/lodestar-beacon-state-transition/src/fast/slot/index.ts
+++ b/packages/lodestar-beacon-state-transition/src/fast/slot/index.ts
@@ -1,12 +1,17 @@
-import {BeaconState, Slot} from "@chainsafe/lodestar-types";
+import {Slot} from "@chainsafe/lodestar-types";
 
 import {StateTransitionEpochContext} from "../util/epochContext";
 import {processEpoch} from "../epoch";
 import {processSlot} from "./processSlot";
+import {CachedValidatorsBeaconState} from "../util/interface";
 
 export {processSlot};
 
-export function processSlots(epochCtx: StateTransitionEpochContext, state: BeaconState, slot: Slot): void {
+export function processSlots(
+  epochCtx: StateTransitionEpochContext,
+  state: CachedValidatorsBeaconState,
+  slot: Slot
+): void {
   if (!(state.slot < slot)) {
     throw new Error("State slot must transition to a future slot: " + `stateSlot=${state.slot} slot=${slot}`);
   }