--- a/packages/lodestar-beacon-state-transition/src/fast/util/index.ts
+++ b/packages/lodestar-beacon-state-transition/src/fast/util/index.ts
@@ -10,6 +10,7 @@ export * from "./epochShuffling";
 export * from "./epochStakeSummary";
 export * from "./flatValidator";
 export * from "./interface";
+export * from "./persistentVector";
 
 /**
  * Exchange Interface of StateContext