--- a/packages/persistent-ts/.babelrc
+++ b/packages/persistent-ts/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "../../.babelrc"
+}