--- a/packages/persistent-ts/src/index.ts
+++ b/packages/persistent-ts/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./List";
+export * from "./Vector";