… packages/lodestar-beacon-state-transition/test/perf/epoch_processing.test.ts
… packages/lodestar-beacon-state-transition/test/perf/sanity/blocks.test.ts
… packages/lodestar-beacon-state-transition/test/perf/sanity/slots.test.ts
… packages/lodestar-beacon-state-transition/test/unit/fast/util/interface.test.ts
… packages/lodestar/src/api/impl/validator/validator.ts
… packages/lodestar/src/chain/blocks/stateTransition.ts
… packages/lodestar/src/chain/chain.ts
… packages/lodestar/src/chain/eventHandlers.ts
… packages/lodestar/src/chain/factory/block/index.ts
… packages/lodestar/src/chain/initState.ts
… packages/lodestar/src/db/api/beacon/stateContextCache.ts
… packages/lodestar/src/tasks/tasks/archiveStates.ts
… packages/lodestar/test/e2e/chain/factory/block/assembleBlock.test.ts
… packages/lodestar/test/unit/api/impl/beacon/state/utils.test.ts
… packages/lodestar/test/unit/api/impl/validator/duties/proposer.test.ts
… packages/lodestar/test/unit/chain/attestation/process.test.ts
… packages/lodestar/test/unit/chain/chain.test.ts
… packages/lodestar/test/unit/chain/factory/block/blockAssembly.test.ts
… packages/lodestar/test/unit/chain/validation/aggregateAndProof.test.ts
… packages/lodestar/test/unit/chain/validation/attestation.test.ts
… packages/lodestar/test/unit/chain/validation/block.test.ts
… packages/lodestar/test/unit/chain/validation/voluntaryExit.test.ts
… packages/lodestar/test/utils/mocks/chain/chain.ts
… packages/lodestar/test/utils/state.ts
… packages/persistent-ts/LICENSE
… packages/persistent-ts/README.md
… packages/persistent-ts/package.json
… packages/persistent-ts/src/List.ts
… packages/persistent-ts/src/Vector.ts
… packages/persistent-ts/test/perf/Vector.test.ts