resolves #1789, resolves #1878, and resolves #1761

+ Create a CachedValidatorsBeaconState that cache validators using persistent vector (see https://github.com/tuyennhv/persistent-ts/pull/1) and use it everywhere
+ Persistent vector is an immutable data structure, similar to the persistent-merkle-tree we used: it shares a lot of data across beacon state since validators are rarely changed and it has cheap `clone()`
+ Persistent vector loop speed is comparable to regular javascript array loop while it's slow if we loop through `TreeBacked<BeaconState>.validators`
+ When we want to mutate validators inside state, use `setValidator()` and `addValidator()` instead of accessing `validators` directly
+ Average epoch transition time in my local environment for Pyrmont sync was reduced from `3.5s-4.2s` to `1.2s-1.6s`. This should speed up the validator exit process too.
## Testing
+ Pyrmont: synced from slot 0 to head and stay synced for more than 4 hours (tested 2 times)
+ Don't see any memory differences during Pyrmont sync
+ Performance tests

|test|master|this branch|
|----|------|-----------|
|prepareEpochProcessState|1500|100|
|process block of 16 validator exits|6000|200|
|transition - 32 slots|2800|1100|
|transition - 64 slots|5300|2200|
|transition - 128 slots|11000|4300|

## TODO
+ Persistent vector was started from https://github.com/cronokirby/persistent-ts, we should consider having a separate repository for it or just have a separate package like in this PR